Completely changed tack. GTK-based UI begins
Ok, so instead of building and entire OpenGL-based UI framework,
which would have been problematic anyway because it would have to run
on the same thread as all the stuff that does complex compute, and would
thus be a little crap.

So now I'm going to build the UI in GTK, in such a way that it stays adaptable
to other frameworks (think ios and android).

Done.